# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scholr — Scholarships & Opportunities Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List/search opportunities (public)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: category
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: fundingType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [published, draft, archived]
        - name: deadline
          in: query
          schema:
            type: string
            enum: [soon, this_month, open]
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, deadline_soonest, most_viewed]
            default: latest
        - name: featured
          in: query
          schema:
            type: boolean
        - name: authorId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunitiesPage"

    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create a new opportunity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
...
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get single opportunity by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update an opportunity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"

    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete an opportunity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /opportunities/slug/{slug}:
    get:
      operationId: getOpportunityBySlug
      tags: [opportunities]
      summary: Get opportunity by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opportunities/{id}/view:
    post:
      operationId: incrementOpportunityViews
      tags: [opportunities]
      summary: Increment view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: View counted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /opportunities/{id}/related:
    get:
      operationId: getRelatedOpportunities
      tags: [opportunities]
      summary: Get related opportunities by category/country
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
  /opportunities/bulk:
    post:
      operationId: bulkUpdateOpportunities
      tags: [opportunities]
      summary: Bulk update status or delete opportunities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateBody"
      responses:
        "200":
          description: Bulk updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ──── USERS / TEAM ────────────────────────────────────────────────────────
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update a team member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Remove a team member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ──── SETTINGS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]
        avatarUrl:
          type: string
          nullable: true
        postsCount:
          type: integer
        lastActive:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, postsCount, createdAt]
...
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]
        avatarUrl:
          type: string
          nullable: true

    # ── Opportunity ───────────────────────────────────────────────────────
    Opportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        fundingType:
          type: string
          nullable: true
        studyLevel:
          type: array
          items:
            type: string
          nullable: true
        deadline:
          type: string
          nullable: true
        amount:
          type: string
          nullable: true
        applyLink:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        featured:
          type: boolean
        pinned:
          type: boolean
        views:
          type: integer
        authorId:
          type: string
          nullable: true
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        seoTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, status, featured, pinned, views, createdAt, updatedAt]
    CreateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        fundingType:
          type: string
          nullable: true
        studyLevel:
          type: array
          items:
            type: string
          nullable: true
        deadline:
          type: string
          nullable: true
        amount:
          type: string
          nullable: true
        applyLink:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
          default: draft
        featured:
          type: boolean
          default: false
        pinned:
          type: boolean
          default: false
        seoTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
      required: [title]
    UpdateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        fundingType:
          type: string
          nullable: true
        studyLevel:
          type: array
          items:
            type: string
          nullable: true
        deadline:
          type: string
          nullable: true
        amount:
          type: string
          nullable: true
        applyLink:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        featured:
          type: boolean
        pinned:
          type: boolean
        seoTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
    BulkUpdateBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        action:
          type: string
          enum: [publish, archive, delete]
      required: [ids, action]
    OpportunitiesPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
...